requirements.txt
Flask
mysql-connector-python
python-dotenv
Werkzeug
Flask-Limiter
Flask-WTF
gunicorn
